Fix additional cases where return type is specified to be 'boolean', but actual return value might be 'nil' #nojira (#363)
### Problem
First problem is with `hasDirectivesInSelection` is defined to return 'boolean' but when 'nestedCheck' is 'nil', return value is actually 'nil' as well.

Second problem is with `function(aConfig: RemoveArgumentsConfig)` function, 'Array.some' accepts a functor that should return 'boolean', but when 'argument.value' is 'nil', return value is 'nil' as well.

### Solution
Transform the 'nil' result into 'false' using `or false`.
In general, when using 'and' operator on values that are not booleans, 'a and b or c'/'if a then b else c' should be used.

Stylua formatting was disabled because auto-formatting breaks the code.

---
* Two more cases where return type is specified to be 'boolean', but actual return value might be 'nil'

* Applied stylua

* Revert "Applied stylua"

This reverts commit b538dc0b89b7e7ebc1b48716d293942aae7dcaea.

* Add stylua ignore directives because it break code during re-format

* Still apply stylua because reasons

* Revert "Add stylua ignore directives because it break code during re-format"

This reverts commit 240465278ce7ef8f5258fb5e1b8c091d6fdd6bdc.

* Apply stylua

* Apply stylua again